add Transactional

+++ src/main/java/gr/uom/opensource/student/StudentServiceDB.java
@@ -4,15 +4,16 @@ import gr.uom.opensource.course.Course;
 import gr.uom.opensource.registration.Registration;
 import gr.uom.opensource.registration.RegistrationRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Transactional;
 
-import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service("StudentServiceDB")
 @Transactional
-public class StudentServiceDB implements StudentService{
+public class StudentServiceDB implements StudentService {
 
     private final StudentRepository studentRepository;
     private final RegistrationRepository registrationRepository;
@@ -33,35 +34,30 @@ public class StudentServiceDB implements StudentService{
         return studentRepository.save(studentToAdd);
     }
 
+    @Transactional(isolation = Isolation.SERIALIZABLE)
     @Override
     public Student getStudentById(int id) {
         Optional<Student> byId = studentRepository.findById(id); // SELECT * FROM student where student.id=?
         return byId.orElse(new Student());
     }
 
+    @Transactional(rollbackFor = RuntimeException.class)
     @Override
     public Student updateStudent(int id, Student updatedStudent) {
-        Optional<Student> byId = studentRepository.findById(id);
+        Optional<Student> studentOptional = studentRepository.findById(id);
 
-        if (byId.isEmpty())
-            return new Student();
+        Student student = studentOptional.orElseThrow(() -> new RuntimeException("Student Does Not Exist"));
 
-        Student student  = byId.get();
-
-        if(student.getId()==id){
-            if(!student.getFirstName().equals(updatedStudent.getFirstName()) && updatedStudent.getFirstName()!=null)
-                student.setFirstName(updatedStudent.getFirstName());
-
-            if (!student.getLastName().equals(updatedStudent.getLastName()) && updatedStudent.getLastName()!=null)
-                student.setLastName(updatedStudent.getLastName());
-
-            return student;
-        }
-        return new Student();
+        student.setFirstName(updatedStudent.getFirstName() != null ? updatedStudent.getFirstName() : student.getFirstName());
+        student.setLastName(updatedStudent.getLastName() != null ? updatedStudent.getLastName() : student.getLastName());
+        studentRepository.save(student);
+        return student;
     }
 
     @Override
     public void deleteStudent(int id) {
+        if (!studentRepository.existsById(id))
+            throw new RuntimeException("Student Not Found");
         studentRepository.deleteById(id);
     }
